python/requirements.txt
# Data processing and analysis
pandas>=1.5.0
numpy>=1.21.0
scikit-learn>=1.0.0
psycopg2-binary>=2.9.0

# Trading and financial
yfinance>=0.2.0
backtesting>=0.3.3
ta>=0.10.0
newsapi-python>=0.2.6

# Deep Learning
tensorflow>=2.8.0
transformers>=4.15.0
torch>=1.10.0

# API and Web
fastapi>=0.68.0
uvicorn>=0.15.0
requests>=2.28.0
python-dotenv>=0.19.0

# Optimization
optuna>=2.10.0

# Visualization
plotly>=5.3.0
matplotlib>=3.4.0
